Added delete entire entry
I added the button with a trash can so users don
't have to keep clicking on the - button to remove a whole stock pile.
I was also able to migrate some of the functions to another file
to make things simpler and easier to update. Now the main file looks cleaner.